# frontend/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Самый Крутой Бэк",
    "description": "Бэкэнд для управления договорами, счетами, платежами и шаблонами",
    "version": "1.0.0"
  },
  "paths": {
    "/contracts/": {
      "post": {
        "tags": ["Contracts"],
        "summary": "Create New Contract",
        "description": "Создать новый контракт.",
        "operationId": "create_new_contract_contracts__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContractCreate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContractResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/contracts/{contract_id}": {
      "get": {
        "tags": ["Contracts"],
        "summary": "Get Contract Detail",
        "description": "Получить детали контракта.",
        "operationId": "get_contract_detail_contracts__contract_id__get",
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Contract Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContractResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/contracts/{contract_id}/addendums/": {
      "post": {
        "tags": ["Addendums"],
        "summary": "Create New Addendum",
        "description": "Создать новое дополнительное соглашение.",
        "operationId": "create_new_addendum_contracts__contract_id__addendums__post",
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Contract Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddendumCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AddendumResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Addendums"],
        "summary": "List Addendums",
        "description": "Получить список дополнительных соглашений.",
        "operationId": "list_addendums_contracts__contract_id__addendums__get",
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Contract Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AddendumResponse" },
                  "title": "Response List Addendums Contracts  Contract Id  Addendums  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/contracts/{contract_id}/addendums/{addendum_id}": {
      "delete": {
        "tags": ["Addendums"],
        "summary": "Remove Addendum",
        "description": "Удалить дополнительное соглашение.",
        "operationId": "remove_addendum_contracts__contract_id__addendums__addendum_id__delete",
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Contract Id" }
          },
          {
            "name": "addendum_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Addendum Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/contracts/{contract_id}/specifications/contract/{specification_id}": {
      "get": {
        "tags": ["Specifications"],
        "summary": "Get Contract By Specification Endpoint",
        "description": "Получить контракт по спецификации.",
        "operationId": "get_contract_by_specification_endpoint_contracts__contract_id__specifications_contract__specification_id__get",
        "parameters": [
          {
            "name": "specification_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Specification Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/exchange-rate/contract/{contract_id}": {
      "put": {
        "tags": ["Exchange Rate"],
        "summary": "Update Contract Rate",
        "description": "Обновить курс валюты в контракте.",
        "operationId": "update_contract_rate_exchange_rate_contract__contract_id__put",
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Contract Id" }
          },
          {
            "name": "new_rate",
            "in": "query",
            "required": true,
            "schema": { "type": "number", "title": "New Rate" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/pdf/contract/{contract_id}": {
      "get": {
        "tags": ["PDF Export"],
        "summary": "Export Contract",
        "description": "Генерация контракта в формате PDF или DOCX.",
        "operationId": "export_contract_pdf_contract__contract_id__get",
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Contract Id" }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Output file format",
              "enum": ["pdf", "docx"],
              "default": "pdf",
              "title": "Format"
            },
            "description": "Output file format"
          },
          {
            "name": "template",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Template file to use",
              "default": "contract_template.docx",
              "title": "Template"
            },
            "description": "Template file to use"
          },
          {
            "name": "exclude_fields",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": { "type": "string" },
              "description": "List of fields to exclude",
              "default": [],
              "title": "Exclude Fields"
            },
            "description": "List of fields to exclude"
          }
        ],
        "responses": {
          "200": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddendumCreate": {
        "properties": {
          "addendum_number": { "type": "string", "title": "Addendum Number" },
          "addendum_date": {
            "type": "string",
            "format": "date",
            "title": "Addendum Date"
          },
          "additional_amount": {
            "anyOf": [{ "type": "number" }, { "type": "string" }],
            "title": "Additional Amount"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          }
        },
        "type": "object",
        "required": ["addendum_number", "addendum_date", "additional_amount"],
        "title": "AddendumCreate"
      },
      "AddendumResponse": {
        "properties": {
          "addendum_number": { "type": "string", "title": "Addendum Number" },
          "addendum_date": {
            "type": "string",
            "format": "date",
            "title": "Addendum Date"
          },
          "additional_amount": {
            "type": "string",
            "title": "Additional Amount"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "id": { "type": "integer", "title": "Id" },
          "contract_id": { "type": "integer", "title": "Contract Id" }
        },
        "type": "object",
        "required": [
          "addendum_number",
          "addendum_date",
          "additional_amount",
          "id",
          "contract_id"
        ],
        "title": "AddendumResponse"
      },
...
      "ContractCreate": {
        "properties": {
          "contract_number": {
            "type": "string",
            "maxLength": 50,
            "title": "Contract Number"
          },
          "contract_date": {
            "type": "string",
            "format": "date",
            "title": "Contract Date"
          },
          "valid_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "Valid Date"
          },
          "place_of_signing": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Place Of Signing"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "CNY"
          },
          "exchange_rate": {
            "anyOf": [
              { "type": "number" },
              { "type": "string" },
              { "type": "null" }
            ],
            "title": "Exchange Rate"
          },
          "supplier_name": {
            "type": "string",
            "maxLength": 255,
            "title": "Supplier Name"
          },
          "supplier_representative": {
            "type": "string",
            "maxLength": 255,
            "title": "Supplier Representative"
          },
          "supplier_address": {
            "type": "string",
            "maxLength": 255,
            "title": "Supplier Address"
          },
          "supplier_bank": {
            "type": "string",
            "maxLength": 255,
            "title": "Supplier Bank"
          },
          "supplier_account": {
            "type": "string",
            "maxLength": 50,
            "title": "Supplier Account"
          },
          "buyer_name": {
            "type": "string",
            "maxLength": 255,
            "title": "Buyer Name"
          },
          "buyer_address": {
            "type": "string",
            "maxLength": 255,
            "title": "Buyer Address"
          },
          "buyer_bank": {
            "type": "string",
            "maxLength": 255,
            "title": "Buyer Bank"
          },
          "buyer_account": {
            "type": "string",
            "maxLength": 50,
            "title": "Buyer Account"
          },
          "goods_name": {
            "type": "string",
            "maxLength": 255,
            "title": "Goods Name"
          },
          "quantity": { "type": "integer", "title": "Quantity" },
          "price_per_unit": {
            "anyOf": [{ "type": "number" }, { "type": "string" }],
            "title": "Price Per Unit"
          },
          "total_price": {
            "anyOf": [{ "type": "number" }, { "type": "string" }],
            "title": "Total Price"
          },
          "payment_date": {
            "type": "string",
            "format": "date",
            "title": "Payment Date"
          },
          "delivery_terms": {
            "type": "string",
            "maxLength": 50,
            "title": "Delivery Terms"
          },
          "claim_period": { "type": "integer", "title": "Claim Period" },
          "response_period": { "type": "integer", "title": "Response Period" }
        },
        "type": "object",
        "required": [
          "contract_number",
          "contract_date",
          "supplier_name",
          "supplier_representative",
          "supplier_address",
          "supplier_bank",
          "supplier_account",
          "buyer_name",
          "buyer_address",
          "buyer_bank",
          "buyer_account",
          "goods_name",
          "quantity",
          "price_per_unit",
          "total_price",
          "payment_date",
          "delivery_terms",
          "claim_period",
          "response_period"
        ],
        "title": "ContractCreate",
        "description": "Использует `ContractBase`, но отдельно определяет ID при создании."
      },
      "ContractResponse": {
        "properties": {
          "contract_number": {
            "type": "string",
            "maxLength": 50,
            "title": "Contract Number"
          },
          "contract_date": {
            "type": "string",
            "format": "date",
            "title": "Contract Date"
          },
          "valid_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "Valid Date"
          },
          "place_of_signing": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Place Of Signing"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "CNY"
          },
          "exchange_rate": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Exchange Rate"
          },
          "supplier_name": {
            "type": "string",
            "maxLength": 255,
            "title": "Supplier Name"
          },
          "supplier_representative": {
            "type": "string",
            "maxLength": 255,
            "title": "Supplier Representative"
          },
          "supplier_address": {
            "type": "string",
            "maxLength": 255,
            "title": "Supplier Address"
          },
          "supplier_bank": {
            "type": "string",
            "maxLength": 255,
            "title": "Supplier Bank"
          },
          "supplier_account": {
            "type": "string",
            "maxLength": 50,
            "title": "Supplier Account"
          },
          "buyer_name": {
            "type": "string",
            "maxLength": 255,
            "title": "Buyer Name"
          },
          "buyer_address": {
            "type": "string",
            "maxLength": 255,
            "title": "Buyer Address"
          },
          "buyer_bank": {
            "type": "string",
            "maxLength": 255,
            "title": "Buyer Bank"
          },
          "buyer_account": {
            "type": "string",
            "maxLength": 50,
            "title": "Buyer Account"
          },
          "goods_name": {
            "type": "string",
            "maxLength": 255,
            "title": "Goods Name"
          },
          "quantity": { "type": "integer", "title": "Quantity" },
          "price_per_unit": { "type": "string", "title": "Price Per Unit" },
          "total_price": { "type": "string", "title": "Total Price" },
          "payment_date": {
            "type": "string",
            "format": "date",
            "title": "Payment Date"
          },
          "delivery_terms": {
            "type": "string",
            "maxLength": 50,
            "title": "Delivery Terms"
          },
          "claim_period": { "type": "integer", "title": "Claim Period" },
          "response_period": { "type": "integer", "title": "Response Period" },
          "id": { "type": "integer", "title": "Id" }
        },
        "type": "object",
        "required": [
          "contract_number",
          "contract_date",
          "supplier_name",
          "supplier_representative",
          "supplier_address",
          "supplier_bank",
          "supplier_account",
          "buyer_name",
          "buyer_address",
          "buyer_bank",
          "buyer_account",
          "goods_name",
          "quantity",
          "price_per_unit",
          "total_price",
          "payment_date",
          "delivery_terms",
          "claim_period",
          "response_period",
          "id"
        ],
        "title": "ContractResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}